Pull some tests out of extremely bloated `modeNormal.test.ts` into dedicated files
