+++ test/search/motionIncSearch.test.ts
@@ -1,6 +1,6 @@
-import { Configuration } from './testConfiguration';
-import { newTest } from './testSimplifier';
-import { cleanUpWorkspace, setupWorkspace } from './testUtils';
+import { Configuration } from '../testConfiguration';
+import { newTest } from '../testSimplifier';
+import { cleanUpWorkspace, setupWorkspace } from '../testUtils';
 
 suite('incsearch motion', () => {
   suiteSetup(async () => {
